refactor: extract file watching module
